Change How Go Compiler Is Used
With K*S 1.26.3/1.25.8/1.24.12 upstream now builds with an explicit version of go and will try to fetch that version if not present. FORCE_HOTS_GO=y overrides that specification by using the host's version of go. This spec file continues to use built requires to require as a minimum the 'built with' go version from upstream.

Also adds missing dotfiles to src directory during build.