Retail Existing Go Policy
With this release (1.24.12, 1.25.8, 1.26.3) upstream changed the use og
Go to explicitly use the go language version specified in the build
process, and, if not available, to download the appropriate version of
go. This commit retains existing Fedora usage of using the version
specified as 'built with' or newer during the build process.

The commit also includes missing go-related dot files during the build
process.