Update to Kubernetes 1.25.8.
Resolves, in part, 2179470. With this release upstream now uses an
explicit version of go during the build and will download that version
if possible rather than use the version present on the build host.
Fedora will continue to use the host go compiler as long as it is the
same as or newer that the 'built with' setting from upstream.